# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: University Course, Fee, Intake & Requirement Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /universities:
    get:
      operationId: listUniversities
      tags: [universities]
      summary: List all universities
      parameters:
        - name: search
          in: query
          schema:
            type: ["string", "null"]
        - name: country
          in: query
          schema:
            type: ["string", "null"]
        - name: city
          in: query
          schema:
            type: ["string", "null"]
        - name: page
          in: query
          schema:
            type: ["integer", "null"]
        - name: limit
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: List of universities
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UniversityListResponse"
    post:
      operationId: createUniversity
      tags: [universities]
      summary: Create a university
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUniversityBody"
      responses:
        "201":
          description: Created university
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/University"
  /universities/{id}:
    get:
      operationId: getUniversity
      tags: [universities]
      summary: Get a university by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: University
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/University"
        "404":
          description: Not found
    patch:
      operationId: updateUniversity
      tags: [universities]
      summary: Update a university
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUniversityBody"
      responses:
        "200":
          description: Updated university
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/University"
        "404":
          description: Not found
    delete:
      operationId: deleteUniversity
      tags: [universities]
      summary: Delete a university
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    University:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        country:
          type: string
        city:
          type: string
        website:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        logoUrl:
          type: ["string", "null"]
        scrapeUrl:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - country
        - city
        - createdAt
        - updatedAt
    UniversityListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/University"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - data
        - total
        - page
        - limit
    CreateUniversityBody:
      type: object
      properties:
        name:
          type: string
        country:
          type: string
        city:
          type: string
        website:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        logoUrl:
          type: ["string", "null"]
        scrapeUrl:
          type: ["string", "null"]
      required:
        - name
        - country
        - city
    UpdateUniversityBody:
      type: object
      properties:
        name:
          type: ["string", "null"]
        country:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        website:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        logoUrl:
          type: ["string", "null"]
        scrapeUrl:
          type: ["string", "null"]
...